extract common system provider
