+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/provider/BaseJpaSystemProvider.java
@@ -52,7 +52,7 @@ import org.springframework.beans.factory.annotation.Required;
 import javax.servlet.http.HttpServletRequest;
 import java.util.Date;
 
-public class BaseJpaSystemProvider<T, MT> extends BaseJpaProvider implements IJpaSystemProvider {
+public class BaseJpaSystemProvider<T, MT> extends BaseStorageSystemProvider<T, MT> implements IJpaSystemProvider {
 	private static final Logger ourLog = LoggerFactory.getLogger(BaseJpaSystemProvider.class);
 
 	public static final String RESP_PARAM_SUCCESS = "success";
@@ -70,7 +70,6 @@ public class BaseJpaSystemProvider<T, MT> extends BaseJpaProvider implements IJp
 	@Deprecated
 	public static final String PERFORM_REINDEXING_PASS = ProviderConstants.PERFORM_REINDEXING_PASS;
 
-	private IFhirSystemDao<T, MT> myDao;
 	@Autowired
 	private IResourceReindexingSvc myResourceReindexingSvc;
 
@@ -84,36 +83,6 @@ public class BaseJpaSystemProvider<T, MT> extends BaseJpaProvider implements IJp
 
 	protected IResourceReindexingSvc getResourceReindexingSvc() {
 		return myResourceReindexingSvc;
-	}
-
-	@Operation(name = ProviderConstants.OPERATION_EXPUNGE, idempotent = false, returnParameters = {
-		@OperationParam(name = JpaConstants.OPERATION_EXPUNGE_OUT_PARAM_EXPUNGE_COUNT, typeName = "integer")
-	})
-	public IBaseParameters expunge(
-		@OperationParam(name = ProviderConstants.OPERATION_EXPUNGE_PARAM_LIMIT, typeName = "integer") IPrimitiveType<Integer> theLimit,
-		@OperationParam(name = ProviderConstants.OPERATION_EXPUNGE_PARAM_EXPUNGE_DELETED_RESOURCES, typeName = "boolean") IPrimitiveType<Boolean> theExpungeDeletedResources,
-		@OperationParam(name = ProviderConstants.OPERATION_EXPUNGE_PARAM_EXPUNGE_PREVIOUS_VERSIONS, typeName = "boolean") IPrimitiveType<Boolean> theExpungeOldVersions,
-		@OperationParam(name = ProviderConstants.OPERATION_EXPUNGE_PARAM_EXPUNGE_EVERYTHING, typeName = "boolean") IPrimitiveType<Boolean> theExpungeEverything,
-		RequestDetails theRequestDetails
-	) {
-		ExpungeOptions options = createExpungeOptions(theLimit, theExpungeDeletedResources, theExpungeOldVersions, theExpungeEverything);
-		ExpungeOutcome outcome = getDao().expunge(options, theRequestDetails);
-		return createExpungeResponse(outcome);
-	}
-
-	protected IBaseParameters doExpunge(IPrimitiveType<? extends Integer> theLimit, IPrimitiveType<? extends Boolean> theExpungeDeletedResources, IPrimitiveType<? extends Boolean> theExpungeOldVersions, IPrimitiveType<? extends Boolean> theExpungeEverything, RequestDetails theRequestDetails) {
-		ExpungeOptions options = createExpungeOptions(theLimit, theExpungeDeletedResources, theExpungeOldVersions, theExpungeEverything);
-		ExpungeOutcome outcome = getDao().expunge(options, theRequestDetails);
-		return createExpungeResponse(outcome);
-	}
-
-	protected IFhirSystemDao<T, MT> getDao() {
-		return myDao;
-	}
-
-	@Required
-	public void setDao(IFhirSystemDao<T, MT> theDao) {
-		myDao = theDao;
 	}
 
 	@History
