+++ hapi-fhir-jpaserver-mdm/src/test/java/ca/uhn/fhir/jpa/mdm/svc/MdmLinkSvcTest.java
@@ -2,6 +2,7 @@ package ca.uhn.fhir.jpa.mdm.svc;
 
 import ca.uhn.fhir.interceptor.model.RequestPartitionId;
 import ca.uhn.fhir.i18n.Msg;
+import ca.uhn.fhir.jpa.dao.expunge.ExpungeEverythingService;
 import ca.uhn.fhir.jpa.dao.expunge.IExpungeEverythingService;
 import ca.uhn.fhir.jpa.entity.MdmLink;
 import ca.uhn.fhir.jpa.mdm.BaseMdmR4Test;
@@ -35,7 +36,7 @@ public class MdmLinkSvcTest extends BaseMdmR4Test {
 	@Autowired
 	IMdmLinkSvc myMdmLinkSvc;
 	@Autowired
-	IExpungeEverythingService myExpungeEverythingService;
+	ExpungeEverythingService myExpungeEverythingService;
 
 	@Override
 	@AfterEach
